+++ src/DepthTensor/_core/random/generator.py
@@ -28,17 +28,19 @@ except (ImportError, ModuleNotFoundError):
 ###
 
 @overload
-def rand(*, device: DeviceLike = "cpu") -> TensorLike: ...
+def rand(*, device: DeviceLike = "cpu", requires_grad: bool = False) -> TensorLike: ...
 @overload
 def rand(
     *d: int,
     dtype: Optional[DTypeLike] = None,
-    device: DeviceLike = "cpu"
+    device: DeviceLike = "cpu",
+    requires_grad: bool = False
 ) -> TensorLike: ... 
 def rand(
     *d: int,
     dtype: Optional[DTypeLike] = None,
-    device: DeviceLike = "cpu"
+    device: DeviceLike = "cpu",
+    requires_grad: bool = False
 ) -> TensorLike:
     from ...tensor import Tensor
     if device == "cpu":
@@ -46,20 +48,22 @@ def rand(
     else:
         if cp is None: raise CuPyNotFound(CUPY_NOT_FOUND_MSG)
         y = cp.random.rand(*d, dtype=dtype)
-    return Tensor(y, dtype=dtype)
+    return Tensor(y, dtype=dtype, requires_grad=requires_grad)
     
 @overload
-def randn(*, device: DeviceLike = "cpu") -> TensorLike: ...
+def randn(*, device: DeviceLike = "cpu", requires_grad: bool = False) -> TensorLike: ...
 @overload
 def randn(
     *d: int,
     dtype: Optional[DTypeLike] = None,
-    device: DeviceLike = "cpu"
+    device: DeviceLike = "cpu",
+    requires_grad: bool = False
 ) -> TensorLike: ... 
 def randn(
     *d: int,
     dtype: Optional[DTypeLike] = None,
-    device: DeviceLike = "cpu"
+    device: DeviceLike = "cpu",
+    requires_grad: bool = False
 ) -> TensorLike:
     from ...tensor import Tensor
     if device == "cpu":
@@ -67,14 +71,15 @@ def randn(
     else:
         if cp is None: raise CuPyNotFound(CUPY_NOT_FOUND_MSG)
         y = cp.random.randn(*d, dtype=dtype)
-    return Tensor(y, dtype=dtype)
+    return Tensor(y, dtype=dtype, requires_grad=requires_grad)
     
 def randint(
     low: int,
     high: Optional[int] = None,
     size: Optional[AxisLike] = None,
     dtype: Any = int64,
-    device: DeviceLike = "cpu"
+    device: DeviceLike = "cpu",
+    requires_grad: bool = False
 ) -> TensorLike:
     from ...tensor import Tensor
     if device == "cpu":
@@ -82,14 +87,15 @@ def randint(
     else:
         if cp is None: raise CuPyNotFound(CUPY_NOT_FOUND_MSG)
         y = cp.random.randint(low=low, high=high, size=size, dtype=dtype)
-    return Tensor(y)
+    return Tensor(y, requires_grad=requires_grad)
 
 def uniform(
     low: float = 0.0,
     high: float = 1.0,
     size: Optional[AxisLike] = None,
     *,
-    device: DeviceLike = "cpu"
+    device: DeviceLike = "cpu",
+    requires_grad: bool = False
 ):
     from ...tensor import Tensor
     if device == "cpu":
@@ -97,7 +103,7 @@ def uniform(
     else:
         if cp is None: raise CuPyNotFound(CUPY_NOT_FOUND_MSG)
         y = cp.random.uniform(low=low, high=high, size=size)
-    return Tensor(y)
+    return Tensor(y, requires_grad=requires_grad)
 
 ###
 ###