+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - 2025-08-26
+
+### Added
+
+- Additional ```requires_grad``` parameter for function ```random.rand()```.
+
+- Additional ```requires_grad``` parameter for function ```random.randn()```.
+
+- Additional ```requires_grad``` parameter for function ```random.randint()```.
+
+- Additional ```requires_grad``` parameter for function ```random.uniform()```.
+
 ## [1.6.0] - 2025-08-25
 
 ### Added